requirements.txt
Django==5.0
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.0
django-cors-headers==4.3.1
coverage==7.3.2